Added clean up on unload event - debugging output
Bug

Each page has a different id, but each are still trying to show ads and
should not be, this is unexpected behavior.  I call this Timer leak. The
object cannot be cleaned up by GC if needed.  In this case I don’t need
the objects to be cleaned up but just stop running.  That also includes
the ad provider controls. Some turn off on unload but some don’t.  This
will also cause a memory leak.

Some time I want the whole adcontrol to be GC'ed as we could create a
new instance on each page load.

Solution

On unload turn off the adproviders timer and also the adcontrol's timer